# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend pendataan perusahaan berbasis kuesioner SE2026 Kuesioner L (UB \u0026 UMKM Bangunan Usaha)",
        "title": "Backend pendataan perusahaan SE2026-L.UB",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/pendataan-umkm": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Pendataan UMKM"
                ],
                "summary": "List semua pendataan UMKM (admin)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PendataanUmkm"
                            }
                        }
                    }
                }
            }
        },
        "/pendataan-umkm/form/{token}": {
            "get": {
                "tags": [
                    "Pendataan UMKM"
                ],
                "summary": "Get info perusahaan UMKM via token (publik)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token 64-char hex",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Perusahaan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pendataan UMKM"
                ],
                "summary": "Submit kuesioner pendataan UMKM (publik)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token 64-char hex",
                        "name": "token",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Kuesioner UMKM data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PendataanUmkm"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.PendataanUmkm"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/pendataan-umkm/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Pendataan UMKM"
                ],
                "summary": "Detail pendataan UMKM (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pendataan UMKM ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PendataanUmkm"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Pendataan": {
            "type": "object",
            "properties": {
                "b1r1": {
                    "type": "string"
                },
                "b1r10a": {
                    "type": "integer"
                },
                "b1r10b": {
                    "type": "integer"
                },
                "b1r11a": {
                    "type": "string"
                },
                "b1r11b": {
                    "type": "string"
                },
                "b1r11c": {
                    "type": "string"
                },
                "b1r11d": {
                    "type": "string"
                },
                "b1r11e": {
                    "type": "string"
                },
                "b1r11f": {
                    "type": "string"
                },
                "b1r12a": {
                    "type": "integer"
                },
                "b1r12b1": {
                    "type": "integer"
                },
                "b1r12b2": {
                    "type": "integer"
                },
                "b1r12b3": {
                    "type": "integer"
                },
                "b1r12b4": {
                    "type": "integer"
                },
                "b1r12b5": {
                    "type": "integer"
                },
                "b1r12b6": {
                    "type": "integer"
                },
                "b1r12c": {
                    "type": "integer"
                },
                "b1r13a": {
                    "type": "integer"
                },
                "b1r13b": {
                    "type": "integer"
                },
                "b1r14": {
                    "type": "integer"
                },
                "b1r15a": {
                    "type": "integer"
                },
                "b1r15b": {
                    "type": "integer"
                },
                "b1r15c": {
                    "type": "integer"
                },
                "b1r16a": {
                    "type": "integer"
                },
                "b1r16b": {
                    "type": "integer"
                },
                "b1r16c": {
                    "type": "integer"
                },
                "b1r17": {
                    "type": "integer"
                },
                "b1r18": {
                    "type": "integer"
                },
                "b1r19a": {
                    "type": "integer"
                },
                "b1r19b": {
                    "type": "integer"
                },
                "b1r2": {
                    "type": "string"
                },
                "b1r20a": {
                    "type": "integer"
                },
                "b1r20b": {
                    "type": "integer"
                },
                "b1r20c": {
                    "type": "integer"
                },
                "b1r21": {
                    "type": "integer"
                },
                "b1r22a": {
                    "type": "number"
                },
                "b1r22b": {
                    "type": "number"
                },
                "b1r22c": {
                    "type": "number"
                },
                "b1r22d": {
                    "type": "number"
                },
                "b1r22e": {
                    "type": "number"
                },
                "b1r22f": {
                    "type": "number"
                },
                "b1r23a": {
                    "type": "number"
                },
                "b1r23b": {
                    "type": "number"
                },
                "b1r23c": {
                    "type": "number"
                },
                "b1r23d": {
                    "type": "number"
                },
                "b1r24a": {
                    "type": "number"
                },
                "b1r24b": {
                    "type": "number"
                },
                "b1r24c": {
                    "type": "number"
                },
                "b1r24c1": {
                    "type": "integer"
                },
                "b1r24d": {
                    "type": "number"
                },
                "b1r25a": {
                    "type": "number"
                },
                "b1r25b": {
                    "type": "number"
                },
                "b1r25c": {
                    "type": "number"
                },
                "b1r25d": {
                    "type": "number"
                },
                "b1r25e": {
                    "type": "number"
                },
                "b1r25f": {
                    "type": "number"
                },
                "b1r25g": {
                    "type": "number"
                },
                "b1r3": {
                    "type": "string"
                },
                "b1r4": {
                    "type": "string"
                },
                "b1r5a": {
                    "type": "string"
                },
                "b1r5b": {
                    "type": "string"
                },
                "b1r5cAlamat": {
                    "type": "string"
                },
                "b1r5cEmail": {
                    "type": "string"
                },
                "b1r5cHomepage": {
                    "type": "string"
                },
                "b1r5cHp": {
                    "type": "string"
                },
                "b1r5cKodePos": {
                    "type": "string"
                },
                "b1r5cRt": {
                    "type": "string"
                },
                "b1r5cRw": {
                    "type": "string"
                },
                "b1r5cTelEkstensi": {
                    "type": "string"
                },
                "b1r5cTelKodeArea": {
                    "type": "string"
                },
                "b1r5cTelNomor": {
                    "type": "string"
                },
                "b1r5d": {
                    "type": "integer"
                },
                "b1r5e": {
                    "type": "string"
                },
                "b1r6a": {
                    "type": "integer"
                },
                "b1r6b": {
                    "type": "string"
                },
                "b1r6c": {
                    "type": "integer"
                },
                "b1r6cLainnya": {
                    "type": "string"
                },
                "b1r7a": {
                    "type": "integer"
                },
                "b1r7b": {
                    "type": "integer"
                },
                "b1r7c": {
                    "type": "integer"
                },
                "b1r7d": {
                    "type": "integer"
                },
                "b1r8a": {
                    "type": "string"
                },
                "b1r8b": {
                    "type": "integer"
                },
                "b1r8c": {
                    "type": "integer"
                },
                "b1r8d": {
                    "type": "string"
                },
                "b1r9a": {
                    "type": "string"
                },
                "b1r9b1": {
                    "type": "integer"
                },
                "b1r9b2": {
                    "type": "integer"
                },
                "b1r9b3": {
                    "type": "integer"
                },
                "b1r9b4": {
                    "type": "integer"
                },
                "b1r9c": {
                    "type": "integer"
                },
                "b1r9d": {
                    "type": "string"
                },
                "b1r9e": {
                    "type": "string"
                },
                "b1r9f": {
                    "type": "string"
                },
                "b1r9g": {
                    "type": "string"
                },
                "b1r9h": {
                    "type": "string"
                },
                "b1r9i": {
                    "type": "integer"
                },
                "b2catatan": {
                    "type": "string"
                },
                "b3r1": {
                    "type": "string"
                },
                "b3r3": {
                    "type": "string"
                },
                "b3r4": {
                    "type": "string"
                },
                "catatanNgibar": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "perusahaan": {
                    "$ref": "#/definitions/models.Perusahaan"
                },
                "perusahaanId": {
                    "type": "integer"
                },
                "submittedAt": {
                    "type": "string"
                },
                "tanggalPelaksanaan": {
                    "type": "string"
                },
                "ttd": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.PendataanUmkm": {
            "type": "object",
            "properties": {
                "b1r1": {
                    "type": "string"
                },
                "b1r10a": {
                    "type": "integer"
                },
                "b1r10b": {
                    "type": "string"
                },
                "b1r10c": {
                    "type": "integer"
                },
                "b1r10cLainnya": {
                    "type": "string"
                },
                "b1r11a": {
                    "type": "string"
                },
                "b1r11b": {
                    "type": "integer"
                },
                "b1r11c": {
                    "type": "integer"
                },
                "b1r11d": {
                    "type": "integer"
                },
                "b1r12a": {
                    "type": "string"
                },
                "b1r12b": {
                    "type": "integer"
                },
                "b1r12c": {
                    "type": "integer"
                },
                "b1r12d": {
                    "type": "string"
                },
                "b1r13a": {
                    "type": "string"
                },
                "b1r13b1": {
                    "type": "integer"
                },
                "b1r13b2": {
                    "type": "integer"
                },
                "b1r13b3": {
                    "type": "integer"
                },
                "b1r13b4": {
                    "type": "integer"
                },
                "b1r13c": {
                    "type": "integer"
                },
                "b1r13d": {
                    "type": "string"
                },
                "b1r13e": {
                    "type": "string"
                },
                "b1r13f": {
                    "type": "string"
                },
                "b1r13g": {
                    "type": "string"
                },
                "b1r13h": {
                    "type": "string"
                },
                "b1r13i": {
                    "type": "integer"
                },
                "b1r14a": {
                    "type": "integer"
                },
                "b1r14b": {
                    "type": "integer"
                },
                "b1r15a": {
                    "type": "string"
                },
                "b1r15b": {
                    "type": "string"
                },
                "b1r15c": {
                    "type": "string"
                },
                "b1r15d": {
                    "type": "string"
                },
                "b1r15e": {
                    "type": "string"
                },
                "b1r15f": {
                    "type": "string"
                },
                "b1r16a": {
                    "type": "integer"
                },
                "b1r16b1": {
                    "type": "integer"
                },
                "b1r16b2": {
                    "type": "integer"
                },
                "b1r16b3": {
                    "type": "integer"
                },
                "b1r16b4": {
                    "type": "integer"
                },
                "b1r16b5": {
                    "type": "integer"
                },
                "b1r16b6": {
                    "type": "integer"
                },
                "b1r16c": {
                    "type": "integer"
                },
                "b1r17a": {
                    "type": "integer"
                },
                "b1r17b": {
                    "type": "integer"
                },
                "b1r18": {
                    "type": "integer"
                },
                "b1r19a": {
                    "type": "integer"
                },
                "b1r19b": {
                    "type": "integer"
                },
                "b1r19c": {
                    "type": "integer"
                },
                "b1r2": {
                    "type": "string"
                },
                "b1r20a": {
                    "type": "integer"
                },
                "b1r20b": {
                    "type": "integer"
                },
                "b1r20c": {
                    "type": "integer"
                },
                "b1r21": {
                    "type": "integer"
                },
                "b1r22": {
                    "type": "integer"
                },
                "b1r23a": {
                    "type": "integer"
                },
                "b1r23b": {
                    "type": "integer"
                },
                "b1r23c": {
                    "type": "integer"
                },
                "b1r24a1": {
                    "type": "integer"
                },
                "b1r24a2": {
                    "type": "integer"
                },
                "b1r24b1": {
                    "type": "integer"
                },
                "b1r24b2": {
                    "type": "integer"
                },
                "b1r24c1": {
                    "type": "integer"
                },
                "b1r24c2": {
                    "type": "integer"
                },
                "b1r25": {
                    "type": "integer"
                },
                "b1r26a": {
                    "type": "number"
                },
                "b1r26b": {
                    "type": "number"
                },
                "b1r26c": {
                    "type": "number"
                },
                "b1r26d": {
                    "type": "number"
                },
                "b1r26e": {
                    "type": "number"
                },
                "b1r26f": {
                    "type": "number"
                },
                "b1r27a": {
                    "type": "number"
                },
                "b1r27b": {
                    "type": "number"
                },
                "b1r27c": {
                    "type": "number"
                },
                "b1r27d": {
                    "type": "number"
                },
                "b1r28a": {
                    "type": "number"
                },
                "b1r28b": {
                    "type": "number"
                },
                "b1r28c": {
                    "type": "number"
                },
                "b1r28c1": {
                    "type": "integer"
                },
                "b1r28d": {
                    "type": "number"
                },
                "b1r29a": {
                    "type": "number"
                },
                "b1r29b": {
                    "type": "number"
                },
                "b1r29c": {
                    "type": "number"
                },
                "b1r29d": {
                    "type": "number"
                },
                "b1r29e": {
                    "type": "number"
                },
                "b1r29f": {
                    "type": "number"
                },
                "b1r29g": {
                    "type": "number"
                },
                "b1r3": {
                    "type": "string"
                },
                "b1r30a": {
                    "type": "number"
                },
                "b1r30b": {
                    "type": "number"
                },
                "b1r30c": {
                    "type": "number"
                },
                "b1r30d": {
                    "type": "number"
                },
                "b1r30e": {
                    "type": "number"
                },
                "b1r30f": {
                    "type": "number"
                },
                "b1r31a": {
                    "type": "number"
                },
                "b1r31b": {
                    "type": "number"
                },
                "b1r31c": {
                    "type": "number"
                },
                "b1r31d": {
                    "type": "number"
                },
                "b1r31e": {
                    "type": "string"
                },
                "b1r32a": {
                    "type": "number"
                },
                "b1r32b": {
                    "type": "number"
                },
                "b1r32c": {
                    "type": "number"
                },
                "b1r32c1": {
                    "type": "integer"
                },
                "b1r32d": {
                    "type": "number"
                },
                "b1r33a": {
                    "type": "number"
                },
                "b1r33b": {
                    "type": "number"
                },
                "b1r33c": {
                    "type": "number"
                },
                "b1r33d": {
                    "type": "number"
                },
                "b1r33e": {
                    "type": "number"
                },
                "b1r33f": {
                    "type": "number"
                },
                "b1r33g": {
                    "type": "number"
                },
                "b1r4": {
                    "type": "string"
                },
                "b1r5": {
                    "type": "string"
                },
                "b1r6": {
                    "type": "string"
                },
                "b1r7": {
                    "type": "string"
                },
                "b1r8a": {
                    "type": "string"
                },
                "b1r8b": {
                    "type": "string"
                },
                "b1r8cAlamat": {
                    "type": "string"
                },
                "b1r8cEmail": {
                    "type": "string"
                },
                "b1r8cHomepage": {
                    "type": "string"
                },
                "b1r8cHp": {
                    "type": "string"
                },
                "b1r8cKodePos": {
                    "type": "string"
                },
                "b1r8cRt": {
                    "type": "string"
                },
                "b1r8cRw": {
                    "type": "string"
                },
                "b1r8cTelEkstensi": {
                    "type": "string"
                },
                "b1r8cTelKodeArea": {
                    "type": "string"
                },
                "b1r8cTelNomor": {
                    "type": "string"
                },
                "b1r8d": {
                    "type": "integer"
                },
                "b1r8e": {
                    "type": "string"
                },
                "b1r9a": {
                    "type": "integer"
                },
                "b1r9b1": {
                    "type": "string"
                },
                "b1r9b2": {
                    "type": "string"
                },
                "b1r9b3": {
                    "type": "string"
                },
                "b2catatan": {
                    "type": "string"
                },
                "b3r1": {
                    "type": "string"
                },
                "b3r3": {
                    "type": "string"
                },
                "b3r4": {
                    "type": "string"
                },
                "catatanNgibar": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "perusahaan": {
                    "$ref": "#/definitions/models.Perusahaan"
                },
                "perusahaanId": {
                    "type": "integer"
                },
                "submittedAt": {
                    "type": "string"
                },
                "tanggalPelaksanaan": {
                    "type": "string"
                },
                "ttd": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Perusahaan": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "b1r1": {
                    "type": "string"
                },
                "b1r2": {
                    "type": "string"
                },
                "b1r3": {
                    "type": "string"
                },
                "b1r4": {
                    "type": "string"
                },
                "contactPerson": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nama": {
                    "type": "string"
                },
                "pendataan": {
                    "$ref": "#/definitions/models.Pendataan"
                },
                "pendataanUmkm": {
                    "$ref": "#/definitions/models.PendataanUmkm"
                },
                "phone": {
                    "type": "string"
                },
                "skalaUsaha": {
                    "$ref": "#/definitions/models.SkalaUsaha"
                },
                "status": {
                    "$ref": "#/definitions/models.StatusPendataan"
                },
                "token": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.SkalaUsaha": {
            "type": "string",
            "enum": [
                "UB",
                "UMKM"
            ],
            "x-enum-varnames": [
                "SkalaUB",
                "SkalaUMKM"
            ]
        },
        "models.StatusPendataan": {
            "type": "string",
            "enum": [
                "BELUM",
                "SELESAI"
            ],
            "x-enum-varnames": [
                "StatusBelum",
                "StatusSelesai"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter the token with the `Bearer ` prefix, e.g. \"Bearer abcde12345\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}